requirements.txt
aniso8601==6.0.0
asn1crypto==0.24.0
astroid==2.2.5
awscli==1.16.145
boto3==1.9.151
botocore==1.12.151
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.1.1
colorama==0.3.9
cryptography==2.6.1
cycler==0.10.0
dask==1.2.2
decorator==4.4.0
docutils==0.14
Flask==1.0.2
flask-marshmallow==0.10.0
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
idna==2.8
isort==4.3.17
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
marshmallow==2.19.2
marshmallow-sqlalchemy==0.16.2
matplotlib==3.1.0
mccabe==0.6.1
networkx==2.3
numpy==1.16.3
pandas==0.24.2
Pillow==6.0.0
pyasn1==0.4.4
pycparser==2.19
pycrypto==2.6.1
pylint==2.3.1
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pyparsing==2.4.0
PySocks==1.6.8
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
PyYAML==3.13
requests==2.21.0
rsa==3.4.2
ruamel.yaml==0.15.46
s3transfer==0.2.0
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.9.0
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.3
toolz==0.9.0
typed-ast==1.3.5
urllib3==1.24.1
virtualenv==16.6.0
Werkzeug==0.15.2
wrapt==1.11.1
